requirements.txt
pytest
allure-pytest
selenium==4.15.2
selene>=2.0.b4